pointer + percent 11/03

+++ libft/ft_count_base.c
@@ -6,13 +6,13 @@
 /*   By: addzikow <addzikow@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/01 12:46:40 by addzikow          #+#    #+#             */
-/*   Updated: 2021/02/01 12:57:45 by addzikow         ###   ########lyon.fr   */
+/*   Updated: 2021/03/11 14:49:06 by addzikow         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t ft_count_base(unsigned int nbr, const char *base)
+size_t ft_count_base(unsigned long nbr, const char *base)
 {
 	size_t i;
 	size_t len_base;
